Refactor : 미사용 코드 제거(member-repository)
