+++ src/main/java/umc/kkijuk/server/member/repository/MemberJpaRepository.java
@@ -10,7 +10,6 @@ import java.util.Optional;
 public interface MemberJpaRepository extends JpaRepository<Member, Long>{
     Optional<Member> findById(Long id);
     Optional<Member> findByEmail(String email);
-    Optional<Member> findByPhoneNumber(String phoneNumber);
     Optional<Member> findBySocialId(String socialId);
     Optional<List<Member>> findByDeleteDateBefore(LocalDate date);
 }
