+++ src/main/java/umc/kkijuk/server/member/repository/MemberRepository.java
@@ -10,8 +10,6 @@ public interface MemberRepository {
     Optional<Member> findById(Long id);
     Optional<Member> findByEmail(String email);
     Member save(Member member);
-    Optional<Member> findByPhoneNumber(String phoneNumber);
-//    Optional<Member> findBySocialId(Long socialId);
     Optional<Member> findBySocialId(String socialId);
     void deleteById(Long id);
     Optional<List<Member>> findByDeleteDateBefore(LocalDate date);
